new file:   coreutils.py 	modified:   phishsim.py
